ddd_project/requirements.txt
django>=4.2
djangoninja>=0.2.0
pydantic>=2.0